chore: improve timings span organization
