+++ packages/rum-core/test/performance-monitoring/capture-navigation.spec.js
@@ -23,15 +23,17 @@
  *
  */
 
+import * as navTiming from '../../src/performance-monitoring/navigation/navigation-timing-spans'
 import {
-  createNavigationTimingSpans,
-  createResourceTimingSpans,
-  createUserTimingSpans,
   captureNavigation,
+  createNavigationTimingSpans,
+  createUserTimingSpans,
+  createResourceTimingSpans,
   getPageLoadMarks
-} from '../../src/performance-monitoring/capture-navigation'
+} from '../../src/performance-monitoring/navigation/capture-navigation'
 import Transaction from '../../src/performance-monitoring/transaction'
 import { PAGE_LOAD, ROUTE_CHANGE } from '../../src/common/constants'
+import { spyOnFunction } from '../../../../dev-utils/jasmine'
 import { extend } from '../../src/common/utils'
 import resourceEntries from '../fixtures/resource-entries'
 import userTimingEntries from '../fixtures/user-timing-entries'
@@ -148,32 +150,6 @@ describe('Capture hard navigation', function () {
       { name: 'Domain lookup', _end: 20, _start: 1 },
       { name: 'Making a connection to the server', _end: 88, _start: 20 },
       { name: 'Fire "load" event', _end: 964, _start: 962 }
-    ])
-  })
-
-  it('should handle a redirection', function () {
-    // Add redirect info
-    timings.redirectStart = 1572362095000
-    timings.redirectEnd = 1572362095181
-    let spans = createNavigationTimingSpans(
-      timings,
-      timings.redirectStart,
-      transactionStart,
-      transactionEnd
-    )
-
-    expect(spans.map(mapSpan)).toEqual([
-      { name: 'Redirect', _end: 181, _start: 0 },
-      { name: 'Domain lookup', _end: 201, _start: 182 },
-      { name: 'Making a connection to the server', _end: 269, _start: 201 },
-      { name: 'Requesting and receiving the document', _end: 390, _start: 270 },
-      {
-        name: 'Parsing the document, executing sync. scripts',
-        _end: 723,
-        _start: 346
-      },
-      { name: 'Fire "DOMContentLoaded" event', _end: 835, _start: 815 },
-      { name: 'Fire "load" event', _end: 1145, _start: 1143 }
     ])
   })
 
@@ -447,6 +423,109 @@ describe('Capture hard navigation', function () {
         domComplete: 962
       })
     )
+  })
+
+  describe('when there is redirection available', () => {
+    it('should create Redirect span when calculating createNavigationTimingSpans', function () {
+      // Add redirect info
+      const timingObj = {
+        ...timings,
+        redirectStart: 1572362095000,
+        redirectEnd: 1572362095181
+      }
+      let spans = createNavigationTimingSpans(
+        timingObj,
+        timingObj.redirectStart,
+        transactionStart,
+        transactionEnd
+      )
+
+      expect(spans.map(mapSpan)).toEqual([
+        { name: 'Redirect', _end: 181, _start: 0 },
+        { name: 'Domain lookup', _end: 201, _start: 182 },
+        { name: 'Making a connection to the server', _end: 269, _start: 201 },
+        {
+          name: 'Requesting and receiving the document',
+          _end: 390,
+          _start: 270
+        },
+        {
+          name: 'Parsing the document, executing sync. scripts',
+          _end: 723,
+          _start: 346
+        },
+        { name: 'Fire "DOMContentLoaded" event', _end: 835, _start: 815 },
+        { name: 'Fire "load" event', _end: 1145, _start: 1143 }
+      ])
+    })
+
+    it('should include the redirect duration when calculating load navigation marks', function () {
+      // Add redirect info
+      const timingObj = {
+        ...timings,
+        redirectStart: 1572362095000,
+        redirectEnd: 1572362095181
+      }
+      const marks = getPageLoadMarks(timingObj)
+      expect(marks.navigationTiming).toEqual(
+        jasmine.objectContaining({
+          responseEnd: 390,
+          domInteractive: 723,
+          domComplete: 1143
+        })
+      )
+    })
+  })
+
+  describe('baseTime when calculating createNavigationTimingSpans', function () {
+    ;[
+      {
+        name:
+          'should use redirectStart as a baseTime when there is redirection',
+        redirectStart: 1,
+        redirectEnd: 2,
+        expected: 'redirectStart'
+      },
+      {
+        name:
+          'should use fetchStart as a baseTime when there is no redirection',
+        redirectStart: 0,
+        redirectEnd: 0,
+        expected: 'fetchStart'
+      }
+    ].forEach(({ name, redirectStart, redirectEnd, expected }) => {
+      it(name, function () {
+        const perfTiming = window.performance.timing
+        const timingsObj = {
+          ...timings,
+          redirectStart,
+          redirectEnd
+        }
+        Object.defineProperty(window.performance, 'timing', {
+          writable: true,
+          value: timingsObj
+        })
+
+        const navTimingSpansSpy = spyOnFunction(
+          navTiming,
+          'createNavigationTimingSpans'
+        ).and.callThrough()
+
+        const tr = new Transaction('test', PAGE_LOAD)
+        tr.captureTimings = true
+        tr.end()
+        captureNavigation(tr)
+
+        expect(navTimingSpansSpy).toHaveBeenCalledWith(
+          timingsObj, // timings
+          timingsObj[expected], // baseTime
+          0, // transaction start
+          tr._end // transaction end
+        )
+
+        window.performance.timing = perfTiming
+      })
+    })
   })
 
   describe('Buggy Navigation Timing data', () => {
